Crashes fixed
TLDR The material instances weren't marked with UPROPERTY and were garbage collected whilst still being used.

This part is mainly me ranting about the pains I've had resolving this issue due to my inexperience with Unreal.
Granted some of this information is useful for the future if I consider multithreading as that was the main part of my testing
The main issue of this error was that it was unrepeatable, it would occur every so often on a paint call, and it would always be on the second usage of a material.
From this I could gather that the material was able to function properly for the start of the function, not to mention the previous times it functioned properly.
This meant somewhere after the first function call of the material something happened causing it to become a nullptr.
So I started looking into where exactly the crashes were happening.
This led me to my main red herring, instance having a value in local.
Within the material instance code, in the GameThread_UpdateMIParameter function, accessing the Instance->GetMaterial_Concurrent() line returned the nullptr.
Because the Instance contained a value within the locals I believed the initialisation of it was correct.
I did some experimenting and changed the order of function calls in an inconsiquential manner.
The errors I recieved from both had functions to do with threads as well as the exception messages mentioning asynchronisation of threads.
From there I believed it was a multithreading issue.
At some point a race condition, to do with the material, lost causing the program to crash.
As later discovered this was not the case but here are the results of my usage of multithreading.
-The entire function cannot be resolved on the game thread as it is dependant on the render thread.
-Because the render thread is slower than the game thread, parts of the function needs to wait for the render thread.
-Not sure why but waiting for the render thread then using AsyncTask on the mesh to return it to its original material doesn't update properly and the material remains as the unwrap.
-A render call was being held back as the scene would wait for the next function call in order to update the material's painted surface.
I think with the way I was handling the threads a lot of the game thread calls were being discarded.
Anyways I wasn't getting any progress done as I had similar results in the same place even with managing the threads
I then read a suggestion of giving my materials a UPROPERTY().
Originally I had discarded this thought as I had a lot of other suggestions as to where my errors were coming from.
The reason for this is my knowledge at the time about UPROPERTY() was telling unreal engine to handle it.
Which means that if it was a simple variable I wouldn't need unreal to handle it.
However I did not realise this also protects it from garbage collection, telling unreal that it's being used and has a reference.
After going through many other simple suggestions I tried the UPROPERTY() as it was quick to test and wouldn't cause any harm.
Here's the tricky part.
My criteria for the crash being fixed was 5 minutes of constant painting without a crash.
And the good news is this simple fix passed that criteria!
There is a small chance that the crash was something else but from the test data, it heavily supports the theory that the crash has been fixed.
Looking back I can see how I should have been hinted at this being the issue with my current knowledge.
But overall, even through the stress, time and energy, I learnt more about unreal engine and as always love coding a little bit more.

I have spent wayyy too long resolving such a small fix so I needed to write down my efforts xd